+++ packages/features/ee/workflows/components/TimeTimeUnitInput.tsx
@@ -17,6 +17,7 @@ const TIME_UNITS = [TimeUnit.DAY, TimeUnit.HOUR, TimeUnit.MINUTE] as const;
 
 type Props = {
   disabled: boolean;
+  defaultTime?: number;
 };
 
 const TimeUnitAddonSuffix = ({
@@ -62,7 +63,7 @@ export const TimeTimeUnitInput = (props: Props) => {
           min="1"
           label=""
           disabled={props.disabled}
-          defaultValue={form.getValues("time") || 24}
+          defaultValue={form.getValues("time") ?? props.defaultTime ?? 24}
           className="-mt-2 rounded-r-none text-sm focus:ring-0"
           {...form.register("time", { valueAsNumber: true })}
           addOnSuffix={
