+++ packages/features/webhooks/components/WebhookForm.tsx
@@ -229,7 +229,7 @@ const WebhookForm = (props: {
         {showTimeSection && (
           <div className="mt-5">
             <Label>{t("how_long_after_user_no_show_minutes")}</Label>
-            <TimeTimeUnitInput disabled={false} />
+            <TimeTimeUnitInput disabled={false} defaultTime={5} />
           </div>
         )}
