+++ packages/features/webhooks/pages/webhook-new-view.tsx
@@ -6,22 +6,12 @@ import { useRouter } from "next/navigation";
 import { useCompatSearchParams } from "@calcom/lib/hooks/useCompatSearchParams";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 import { trpc } from "@calcom/trpc/react";
-import { showToast, SkeletonContainer, SkeletonText } from "@calcom/ui";
+import { showToast } from "@calcom/ui";
 
 import type { WebhookFormSubmitData } from "../components/WebhookForm";
 import WebhookForm from "../components/WebhookForm";
 import { subscriberUrlReserved } from "../lib/subscriberUrlReserved";
 
-const SkeletonLoader = () => {
-  return (
-    <SkeletonContainer>
-      <div className="divide-subtle border-subtle space-y-6 rounded-b-lg border border-t-0 px-6 py-4">
-        <SkeletonText className="h-8 w-full" />
-        <SkeletonText className="h-8 w-full" />
-      </div>
-    </SkeletonContainer>
-  );
-};
 export const NewWebhookView = () => {
   const searchParams = useCompatSearchParams();
   const { t } = useLocale();
@@ -87,7 +77,7 @@ export const NewWebhookView = () => {
     });
   };
 
-  if (isPending) return <SkeletonLoader />;
+  if (isPending) return null;
 
   return (
     <WebhookForm
